Use either tz for date display depending on locale
Instead of using the timezone of the event
use either UTC or CEST for displaying datetimes
on the website as well as in mails
depending on the current locale (website) or the language of the event.

Also default to either UTC or CEST for timezone entry
when creating an event.

fix #35